Make some updates during early SBND data-taking

+++ Xporter/ManualFTSFileCleanup.py
@@ -2,11 +2,17 @@ import subprocess
 import json
 import glob
 import os
+from datetime import datetime
 
 import sys
 
-root_file_list = glob.glob("/data/fts_dropbox/*.root")
-json_file_list = glob.glob("/data/fts_dropbox/*.json")
+root_file_list = glob.glob("/data/sbndraw/fts_dropbox/*.root")
+json_file_list = glob.glob("/data/sbndraw/fts_dropbox/*.json")
+json_file_list.sort(key=lambda x: os.path.getmtime(x))
+
+if not root_file_list:
+    print('No files found. Are you running this program from the event builder machine?')
+nfile = len(json_file_list)
 
 #FTS output string formatted as below
 #${sbn_dm.detector}/${file_type}/${data_tier}/${data_stream}/${icarus_project.version}/${icarus_project.name}/${icarus_project.stage}/${run_number[8/2]}
@@ -14,14 +20,18 @@ json_file_list = glob.glob("/data/fts_dropbox/*.json")
 MATCH_STRING = "NEARLINE"
 if (len(sys.argv) > 1):
     MATCH_STRING = sys.argv[1]
+
+now = datetime.now()
 print(f'Remove files containing {MATCH_STRING} in curl output.')
+print(f'Found {nfile} metadata files in dropbox area')
 
 
-addr_lead_str="https://fndca3b.fnal.gov:3880/api/v1/namespace/pnfs/fnal.gov/usr/icarus/archive/sbn"
+addr_lead_str="https://fndca3b.fnal.gov:3880/api/v1/namespace/pnfs/fnal.gov/usr/sbnd/archive/sbn"
 
-for fname in json_file_list:
-    print(fname)
+for i, fname in enumerate(json_file_list):
     metadata = {}
+    if i % 100 == 0:
+        print(f' checking file {i}/{nfile}')
     with open(fname) as f:
         try:
             metadata = json.load(f)
@@ -30,13 +40,14 @@ for fname in json_file_list:
     if len(metadata)==0:
         continue
     run_number = int(metadata["runs"][0][0])
+    # print(fname)
     append_str = "%s/%s/%s/%s/%s/%s/%s/%02d/%02d/%02d/%02d"%(metadata["sbn_dm.detector"],
                                                                  metadata["file_type"],
                                                                  metadata["data_tier"],
                                                                  metadata["data_stream"],
-                                                                 metadata["icarus_project.version"],
-                                                                 metadata["icarus_project.name"],
-                                                                 metadata["icarus_project.stage"],
+                                                                 metadata["sbnd_project.version"],
+                                                                 metadata["sbnd_project.name"],
+                                                                 metadata["sbnd_project.stage"],
                                                                  run_number//100//100//100%100,
                                                                  run_number//100//100%100,
                                                                  run_number//100%100,
@@ -47,7 +58,7 @@ for fname in json_file_list:
     #print(root_fname)
 
     addr = "%s/%s/%s?locality=true"%(addr_lead_str,append_str,root_fname.split("/")[-1])
-    #print(addr)
+    print(addr)
 
     p = subprocess.Popen(['curl','-k','-X','GET','%s'%addr],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     stdout, stderr = p.communicate()
@@ -62,7 +73,10 @@ for fname in json_file_list:
         print("FILE %s not declared?"%root_fname)
         continue
 
-    print(curl_out["fileLocality"])
+    ts = datetime.fromtimestamp(curl_out['creationTime'] / 1e3)
+    elapsed_days = (now - ts).days
+    print(f'Got file ({elapsed_days:.1f} days old)')
+    print(curl_out)
 
     if "NONE" in curl_out["fileLocality"]:
         print("FILE %s no locality?"%root_fname)
@@ -70,9 +84,13 @@ for fname in json_file_list:
 
     if MATCH_STRING in curl_out["fileLocality"] :
         #print("FILE %s ON TAPE!"%root_fname)
-        print("FILE:", root_fname, "STATUS:", curl_out["fileLocality"], "MATCHES THE REQUIRED PATTERN:", MATCH_STRING, "!")
-        print("Delete files %s and %s"%(fname,root_fname))
-        os.system("rm -f %s"%root_fname)
-        os.system("rm -f %s"%fname)
+        #print("FILE:", root_fname, "STATUS:", curl_out["fileLocality"], "MATCHES THE REQUIRED PATTERN:", MATCH_STRING, "!")
+        print(ts)
+        if elapsed_days > 7:
+            print("Delete files %s and %s (%.1f days old)"%(fname,root_fname,elapsed_days))
+            os.system("rm -f %s"%root_fname)
+            os.system("rm -f %s"%fname)
+    else:
+        print("File not on tape yet, not deleting!")
 
 print("Done")
